pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "sourdough-monitor"
version = "2.0.0"
description = "Automated sourdough fermentation monitoring with AI vision analysis"
requires-python = ">=3.9"
dependencies = [
    "matplotlib>=3.8",
    "requests>=2.31",
    "firebase-admin>=6.5",
    "google-api-python-client>=2.120",
    "google-auth-oauthlib>=1.2",
    "google-auth-httplib2>=0.2",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
cv = ["opencv-python>=4.8", "numpy>=1.24"]
dev = ["pytest>=7.0"]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
